web-app/back-end/requirements.txt
sqlalchemy
fastapi
pydantic
typing
uvicorn
python-dotenv
pymysql
